+++ Changes.md
@@ -1,3 +1,4 @@
+* #1096 - Allow one to optimize tables in groups
 * #1100 - Switch the scoreboard to web sockets to avoid issues with web servers
 * #1091 - Use web sockets for the display updates
 * #1101 - Ensure that enumerated values have distinct scores